requirements.txt
llama-index
llama-index-core
llama-index-embeddings-langchain
llama-index-embeddings-openai
llama-index-legacy
llama-index-llms-huggingface
langchain
langchain-community
langchain-core
pyvis
ipython
pypdf

# Visual RAG dependencies
pillow
pytesseract
torch
torchvision
sentence-transformers
transformers
opencv-python

# Web UI dependencies
streamlit
fastapi
uvicorn
networkx
plotly
streamlit-agraph
